new scaffolding for react components
