+++ _footer.html
@@ -37,11 +37,5 @@
 	</div>
 
 
-
-	<script src="https://code.jquery.com/jquery-2.2.3.min.js"   
-	integrity="sha256-a23g1Nt4dtEYOj7bR+vTu7+T8VP13humZFBJNIYoEJo=" crossorigin="anonymous"></script>
-	
-	<script src="./dist/libs/remodal/dist/remodal.min.js"></script>
-
 	<!-- react -->
 	<script type="text/javascript" src="./dist/js/common.min.js"></script>
