+++ package.json
@@ -5,7 +5,8 @@
   "main": "webpack.config.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack --progress --colors --watch"
+    "start": "webpack --progress --colors --watch",
+    "start": "babel-node server.js"
   },
   "author": "Vlad Feninets",
   "license": "ISC",
@@ -14,8 +15,14 @@
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
+    "bower": "^1.7.9",
+    "bower-webpack-plugin": "^0.1.9",
+    "css-loader": "^0.23.1",
+    "file-loader": "^0.8.5",
+    "path": "^0.12.7",
     "react": "^15.0.1",
     "react-dom": "^15.0.1",
+    "style-loader": "^0.13.1",
     "webpack": "^1.13.0",
     "webpack-dev-server": "^1.14.1"
   }
